backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
mysqlclient==2.2.7
python-decouple==3.8
sqlparse==0.5.3
